making edits to training so that it is wandb compatible
